Rename: 컴포넌트끼리 각 폴더 만들어 이동
